refactor : 미사용 파일 제거 및 파일 명명 규칙 통일화 작업
- 추가 누락 파일 커밋
