--- be/infrastructure/repository/AIModel.ts
@@ -1,3 +1,0 @@
-import { AIModelRepository } from '@/be/domain/chat/AIModelRepository';
-
-export class AIModelRepositoryImpl implements AIModelRepository {}
