+++ be/infrastructure/repository/AIModelRepository.ts
@@ -1,0 +1,24 @@
+import { AIModelRepository } from '@/be/domain/chat/AIModelRepository';
+import { PrismaClient, AIModel } from '@prisma/client';
+
+export class AIModelRepositoryImpl implements AIModelRepository {
+   private prisma: PrismaClient;
+
+   constructor() {
+      this.prisma = new PrismaClient();
+   }
+
+   async getModelById(modelId: string): Promise<AIModel | null> {
+      try {
+         const model = await this.prisma.aiModel.findUnique({
+            where: { id: modelId },
+         });
+         return model;
+      } catch (error) {
+         console.error('Error fetching AI model:', error);
+         throw new Error('Could not fetch AI model');
+      }
+   }
+
+   // 추가적인 메서드 구현...
+}
